mint command relocation
