+++ command/root/root.go
@@ -11,11 +11,11 @@ import (
 	"github.com/0xPolygon/polygon-edge/command/genesis"
 	"github.com/0xPolygon/polygon-edge/command/helper"
 	"github.com/0xPolygon/polygon-edge/command/license"
+	"github.com/0xPolygon/polygon-edge/command/mint"
 	"github.com/0xPolygon/polygon-edge/command/monitor"
 	"github.com/0xPolygon/polygon-edge/command/peers"
 	"github.com/0xPolygon/polygon-edge/command/polybft"
 	"github.com/0xPolygon/polygon-edge/command/regenesis"
-	"github.com/0xPolygon/polygon-edge/command/rootchain"
 	"github.com/0xPolygon/polygon-edge/command/secrets"
 	polybftsecrets "github.com/0xPolygon/polygon-edge/command/secrets/init"
 	"github.com/0xPolygon/polygon-edge/command/server"
@@ -49,7 +49,6 @@ func (rc *RootCommand) registerSubCommands() {
 		status.GetCommand(),
 		secrets.GetCommand(),
 		peers.GetCommand(),
-		rootchain.GetCommand(),
 		monitor.GetCommand(),
 		backup.GetCommand(),
 		genesis.GetCommand(),
@@ -59,6 +58,7 @@ func (rc *RootCommand) registerSubCommands() {
 		polybft.GetCommand(),
 		bridge.GetCommand(),
 		regenesis.GetCommand(),
+		mint.GetCommand(),
 	)
 }
