--- command/rootchain/premine/premine.go
@@ -1,135 +1,0 @@
-package premine
-
-import (
-	"fmt"
-	"time"
-
-	"github.com/0xPolygon/polygon-edge/command"
-	bridgeHelper "github.com/0xPolygon/polygon-edge/command/bridge/helper"
-	"github.com/0xPolygon/polygon-edge/command/helper"
-	polybftsecrets "github.com/0xPolygon/polygon-edge/command/secrets/init"
-	sidechainHelper "github.com/0xPolygon/polygon-edge/command/sidechain"
-	"github.com/0xPolygon/polygon-edge/txrelayer"
-	"github.com/0xPolygon/polygon-edge/types"
-	"github.com/spf13/cobra"
-)
-
-var (
-	params premineParams
-)
-
-func GetCommand() *cobra.Command {
-	premineCmd := &cobra.Command{
-		Use: "premine",
-		Short: "Premine native root token to the caller, which determines genesis balances. " +
-			"This command is used in case Supernets native token is rootchain originated.",
-		PreRunE: runPreRun,
-		RunE:    runCommand,
-	}
-
-	helper.RegisterJSONRPCFlag(premineCmd)
-	setFlags(premineCmd)
-
-	return premineCmd
-}
-
-func setFlags(cmd *cobra.Command) {
-	cmd.Flags().StringVar(
-		&params.accountDir,
-		polybftsecrets.AccountDirFlag,
-		"",
-		polybftsecrets.AccountDirFlagDesc,
-	)
-
-	cmd.Flags().StringVar(
-		&params.accountConfig,
-		polybftsecrets.AccountConfigFlag,
-		"",
-		polybftsecrets.AccountConfigFlagDesc,
-	)
-
-	cmd.Flags().StringVar(
-		&params.privateKey,
-		polybftsecrets.PrivateKeyFlag,
-		"",
-		polybftsecrets.PrivateKeyFlagDesc,
-	)
-
-	cmd.Flags().StringVar(
-		&params.customSupernetManager,
-		bridgeHelper.SupernetManagerFlag,
-		"",
-		bridgeHelper.SupernetManagerFlagDesc,
-	)
-
-	cmd.Flags().StringVar(
-		&params.nativeTokenRoot,
-		bridgeHelper.Erc20TokenFlag,
-		"",
-		"address of root erc20 native token",
-	)
-
-	cmd.Flags().StringVar(
-		&params.rootERC20Predicate,
-		rootERC20PredicateFlag,
-		"",
-		"address of root erc20 predicate",
-	)
-
-	cmd.Flags().StringVar(
-		&params.amount,
-		sidechainHelper.AmountFlag,
-		"",
-		"amount to premine",
-	)
-
-	cmd.MarkFlagsMutuallyExclusive(polybftsecrets.AccountDirFlag, polybftsecrets.AccountConfigFlag)
-	cmd.MarkFlagsMutuallyExclusive(polybftsecrets.PrivateKeyFlag, polybftsecrets.AccountConfigFlag)
-	cmd.MarkFlagsMutuallyExclusive(polybftsecrets.PrivateKeyFlag, polybftsecrets.AccountDirFlag)
-}
-
-func runPreRun(cmd *cobra.Command, _ []string) error {
-	params.jsonRPC = helper.GetJSONRPCAddress(cmd)
-
-	return params.validateFlags()
-}
-
-func runCommand(cmd *cobra.Command, _ []string) error {
-	outputter := command.InitializeOutputter(cmd)
-	defer outputter.WriteOutput()
-
-	ownerKey, err := bridgeHelper.GetECDSAKey(params.privateKey, params.accountDir, params.accountConfig)
-	if err != nil {
-		return err
-	}
-
-	txRelayer, err := txrelayer.NewTxRelayer(txrelayer.WithIPAddress(params.jsonRPC),
-		txrelayer.WithReceiptTimeout(150*time.Millisecond))
-	if err != nil {
-		return err
-	}
-
-	approveTxn, err := bridgeHelper.CreateApproveERC20Txn(
-		params.amountValue,
-		types.StringToAddress(params.customSupernetManager),
-		types.StringToAddress(params.nativeTokenRoot), true)
-	if err != nil {
-		return err
-	}
-
-	receipt, err := txRelayer.SendTransaction(approveTxn, ownerKey)
-	if err != nil {
-		return err
-	}
-
-	if receipt == nil || receipt.Status == uint64(types.ReceiptFailed) {
-		return fmt.Errorf("approve transaction failed on block %d", receipt.BlockNumber)
-	}
-
-	outputter.WriteCommandResult(&premineResult{
-		Address: ownerKey.Address().String(),
-		Amount:  params.amountValue,
-	})
-
-	return nil
-}
