+++ e2e-polybft/framework/test-bridge.go
@@ -50,7 +50,7 @@ func NewTestBridge(t *testing.T, clusterConfig *TestClusterConfig) (*TestBridge,
 func (t *TestBridge) Start() error {
 	// Build arguments
 	args := []string{
-		"rootchain",
+		"bridge",
 		"server",
 		"--data-dir", t.clusterConfig.Dir("test-rootchain"),
 	}
@@ -313,7 +313,7 @@ func (t *TestBridge) deployRootchainContracts(genesisPath string) error {
 	}
 
 	args := []string{
-		"rootchain",
+		"bridge",
 		"deploy",
 		"--stake-manager", polybftConfig.Bridge.StakeManagerAddr.String(),
 		"--stake-token", polybftConfig.Bridge.StakeTokenAddr.String(),
@@ -359,7 +359,7 @@ func (t *TestBridge) fundAddressesOnRoot(tokenConfig *polybft.TokenConfig, polyb
 	// non-validator addresses don't need to mint stake token,
 	// they only need to be funded with root token
 	args := []string{
-		"rootchain",
+		"bridge",
 		"fund",
 	}
 
@@ -513,7 +513,7 @@ func (t *TestBridge) FundValidators(tokenAddress types.Address, secretsPaths []s
 	}
 
 	args := []string{
-		"rootchain",
+		"bridge",
 		"fund",
 		"--stake-token", tokenAddress.String(),
 		"--mint",
@@ -611,7 +611,7 @@ func (t *TestBridge) premineNativeRootToken(tokenConfig *polybft.TokenConfig,
 
 	premineCmdArgs := func(secret, key string, amount *big.Int) error {
 		args := []string{
-			"rootchain",
+			"bridge",
 			"premine",
 			"--jsonrpc", t.JSONRPCAddr(),
 			"--supernet-manager", polybftConfig.Bridge.CustomSupernetManagerAddr.String(),
